@@ -1,119 +1,106 @@
 {
   "name": "igo",
-  "version": "0.19.4",
+  "version": "0.20.0",
   "description": "Infrastructure géomatique ouverte",
   "repository": {
     "url": "https://github.com/infra-geo-ouverte/igo2"
   },
   "license": "LiLiQ-R",
-  "angular-cli": {},
   "scripts": {
     "ng": "ng",
     "start": "ng serve --host 0.0.0.0 --proxy-config proxy.conf.json",
-    "start.dev": "npm start",
-    "start.prod": "ng serve --aot --prod --host 0.0.0.0 --proxy-config proxy.conf.json",
-    "build.dev.jit": "ng build --dev --output-path ./dist/dev",
-    "build.dev": "ng build --output-path ./dist/dev --dev --bh /",
-    "build.prod.jit": "ng build --prod --output-path ./dist/prod",
-    "prebuild.prod": "mkdir -p ./dist",
-    "build.prod": "ng build --output-path ./dist/prod --prod --bh /",
-    "serve.dev": "node_modules/live-server/live-server.js ./dist/dev/ --port=4200 --no-browser",
-    "serve.prod": "node_modules/live-server/live-server.js ./dist/prod/ --port=4200 --no-browser",
+    "build": "ng build",
+    "lint": "ng lint",
+    "e2e": "ng e2e --port 4300",
+    "postinstall": "gulp geo:fixOL",
+    "build.prod": "ng build --prod",
+    "serve.prod": "node_modules/live-server/live-server.js ./dist/igo2/ --port=4200 --no-browser",
     "doc": "compodoc -p src/tsconfig.app.json -s --port 4220",
     "build.doc": "compodoc -p src/tsconfig.app.json",
     "serve.doc": "compodoc -s --port 4220",
     "coverage": "npm run build.coverage && npm run serve.coverage",
-    "build.coverage": "ng test --code-coverage --single-run",
+    "build.coverage": "ng test --code-coverage --watch=false",
     "serve.coverage": "node_modules/live-server/live-server.js ./coverage/ --port=4210 --no-browser",
     "i18n": "ng2-translate-extract --dir ./src --output ./src/assets/locale/ --format=json --clean",
-    "lint": "tslint \"src/**/*.ts\" --project src/tsconfig.app.json && tslint \"e2e/**/*.ts\" --project e2e/tsconfig.e2e.json",
     "pretest": "ng lint",
-    "test": "ng test --single-run",
+    "test": "ng test --watch=false",
     "test.watch": "ng test",
-    "pree2e": "webdriver-manager clean && webdriver-manager update --standalone false --gecko false",
-    "e2e": "protractor",
-    "tests.all": "npm run test",
-    "outdated": "ncu",
-    "outdated.wizard": "npm-check -u",
-    "audit-nsp": "nsp check package.json",
-    "audit": "snyk test",
-    "audit.wizard": "snyk wizard",
-    "audit.protect": "snyk protect",
-    "snyk-protect": "snyk protect",
-    "prepublish": "npm run snyk-protect",
-    "cli-update": "npm cache clean && npm install --save-dev angular-cli@latest && npm run reinstall && ng update --routing --style=styl --prefix=igo",
-    "reinstall": "npm run clean && npm cache clean && npm prune && npm install",
-    "clean": "rimraf ./dist",
+    "test.all": "npm run test && npm run e2e",
+    "clean": "rimraf ./dist && rimraf ./node_modules && npm cache clean --force && git checkout -- package.json && git checkout -- package-lock.json",
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s -r 0",
     "changelog-current": "conventional-changelog -p angular -r 2",
-    "preversion": "npm run tests.all && npm run clean && npm run build.prod",
+    "preversion": "npm run test.all && rimraf ./dist && npm run build.prod",
     "version": "npm run changelog && git add -A CHANGELOG.md",
-    "postversion": "git push && git push --tags && rimraf igo2.zip && cd ./dist/prod/ && zip -r ../igo2.zip ./* && npm run ghpages && npm run ghrelease",
-    "preghpages": "npm run build.prod -- --output-path ./dist/ghpages --bh /igo2/ && echo \"include: ['_default.json', '_contexts.json']\" > dist/ghpages/_config.yml",
-    "ghpages": "VERSION=$(node -p -e \"require('./package.json').version\") && ngh --dir=dist/ghpages --no-silent=false --message=$VERSION",
+    "postversion": "git push && git push --tags && rimraf igo2.zip && cd ./dist/igo2/ && zip -r ../igo2.zip ./* && npm run ghpages && npm run ghrelease",
+    "preghpages": "npm run build.prod -- --output-path ./dist/ghpages --base-href /igo2/ && echo \"include: ['_default.json', '_contexts.json']\" > dist/ghpages/_config.yml",
+    "ghpages": "VERSION=$(node -p -e \"require('./package.json').version\") && npx ngh --dir=dist/ghpages --no-silent=false --message=$VERSION",
     "postghpages": "rimraf ./dist/ghpages",
     "ghrelease": "VERSION=$(node -p -e \"require('./package.json').version\") && publish-release --owner infra-geo-ouverte --repo igo2 --tag $VERSION --name $VERSION --reuse-release --notes \"$(npm run changelog-current -s)\" --assets dist/igo2.zip"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.1.0",
-    "@angular/cdk": "^5.1.0",
-    "@angular/common": "^5.1.0",
-    "@angular/compiler": "^5.1.0",
-    "@angular/core": "^5.1.0",
-    "@angular/forms": "^5.1.0",
-    "@angular/http": "^5.1.0",
-    "@angular/material": "^5.1.0",
-    "@angular/platform-browser": "^5.1.0",
-    "@angular/platform-browser-dynamic": "^5.1.0",
-    "@angular/router": "^5.1.0",
-    "@ngrx/core": "^1.2.0",
-    "@ngrx/store": "^2.2.1",
-    "angular2-notifications": "^1.0.1",
+    "@angular/animations": "^6.0.0",
+    "@angular/cdk": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/forms": "^6.0.0",
+    "@angular/material": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "@angular/router": "^6.0.0",
+    "@igo2/auth": "^0.23.1",
+    "@igo2/common": "^0.23.1",
+    "@igo2/core": "^0.23.1",
+    "@igo2/geo": "^0.23.1",
+    "@igo2/context": "^0.23.1",
+    "@igo2/tools": "^0.23.1",
+    "@igo2/utils": "^0.23.1",
     "classlist.js": "^1.1.20150312",
-    "core-js": "^2.4.1",
+    "core-js": "^2.5.4",
     "hammerjs": "^2.0.8",
-    "@igo2/igo2": "^0.19.10",
-    "intl": "^1.2.5",
-    "jspdf": "^1.3.5",
-    "ng2-translate": "^5.0.0",
-    "openlayers": "^4.6.5",
-    "proj4": "^2.4.3",
-    "rxjs": "^5.5.2",
-    "snyk": "^1.30.1",
-    "ts-md5": "^1.2.0",
+    "jwt-decode": "^2.2.0",
+    "ol": "^5.1.3",
+    "proj4": "^2.4.4",
+    "rxjs": "^6.1.0",
     "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.19"
+    "file-saver": "^1.3.8",
+    "html2canvas": "^1.0.0-alpha.12",
+    "jspdf": "^1.4.1",
+    "jszip": "^3.1.5",
+    "moment": "^2.22.2",
+    "stream": "^0.0.2",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.5",
-    "@angular/compiler-cli": "^5.1.0",
-    "@biesbjerg/ngx-translate-extract": "^2.2.2",
-    "@types/jasmine": "2.8.3",
-    "@types/jasminewd2": "~2.0.2",
+    "@angular-devkit/build-angular": "~0.6.1",
+    "@angular-devkit/build-ng-packagr": "~0.6.1",
+    "@angular/cli": "~6.0.1",
+    "@angular/compiler-cli": "^6.0.0",
     "@types/geojson": "^7946.0.3",
-    "@types/node": "^6.0.68",
-    "@types/openlayers": "^4.6.5",
-    "angular-cli-ghpages": "^0.5.0",
-    "codelyzer": "^4.0.1",
-    "compodoc": "0.0.41",
-    "conventional-changelog-cli": "^1.3.1",
-    "jasmine-core": "~2.8.0",
+    "@types/hammerjs": "^2.0.35",
+    "@types/jasmine": "~2.8.6",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.2.1",
+    "conventional-changelog-cli": "^2.0.1",
+    "gulp": "^3.9.1",
+    "gulp-replace": "^1.0.0",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~1.7.1",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "~1.4.2",
+    "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "npm-check": "^5.4.0",
-    "nsp": "^2.6.3",
-    "protractor": "~5.1.2",
-    "publish-release": "^1.3.3",
-    "rimraf": "^2.6.1",
-    "ts-node": "~3.2.0",
+    "ng-packagr": "^3.0.0-rc.2",
+    "protractor": "~5.3.0",
+    "publish-release": "^1.6.0",
+    "scss-bundle": "^2.3.2",
+    "ts-node": "~5.0.1",
+    "tsickle": ">=0.25.5",
+    "tslib": "^1.7.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
-  },
-  "snyk": true
+    "typescript": "~2.7.2"
+  }
 }
