@@ -1,27 +1,36 @@
 import { Component, OnInit, OnDestroy, ChangeDetectorRef } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { Subscription } from 'rxjs/Subscription';
-import { debounceTime} from 'rxjs/operators';
+import { Subscription } from 'rxjs';
+import { debounceTime } from 'rxjs/operators';
 
-import { AuthService, Context, ContextService, DataSourceService, Feature,
-         FeatureService, IgoMap, LayerService, MapService, MediaService,
-         OverlayService, SearchService, ToolService } from '@igo2/igo2';
+import { MediaService } from '@igo2/core';
+import { AuthService } from '@igo2/auth';
+import { Context, ContextService, ToolService } from '@igo2/context';
+import {
+  DataSourceService,
+  Feature,
+  FeatureService,
+  IgoMap,
+  LayerService,
+  MapService,
+  OverlayService,
+  SearchService
+} from '@igo2/geo';
 
-import { controlSlideX, controlSlideY, mapSlideX, mapSlideY } from './portal.animation';
+import {
+  controlSlideX,
+  controlSlideY,
+  mapSlideX,
+  mapSlideY
+} from './portal.animation';
 
 @Component({
   selector: 'app-portal',
   templateUrl: './portal.component.html',
-  styleUrls: ['./portal.component.styl'],
-  animations: [
-    controlSlideX(),
-    controlSlideY(),
-    mapSlideX(),
-    mapSlideY()
-  ]
+  styleUrls: ['./portal.component.scss'],
+  animations: [controlSlideX(), controlSlideY(), mapSlideX(), mapSlideY()]
 })
 export class PortalComponent implements OnInit, OnDestroy {
-
   static SWIPE_ACTION = {
     RIGHT: 'swiperight',
     LEFT: 'swipeleft'
@@ -40,40 +49,46 @@ export class PortalComponent implements OnInit, OnDestroy {
     }
   });
 
-  public sidenavOpened: boolean = false;
-  public toastOpened: boolean = false;
-  public toastShown: boolean = false;
+  public sidenavOpened = false;
+  public toastOpened = false;
+  public toastShown = false;
 
   // True after the initial context is loaded
   private contextLoaded = false;
 
-  constructor(private route: ActivatedRoute,
-              public authService: AuthService,
-              public featureService: FeatureService,
-              public mediaService: MediaService,
-              public toolService: ToolService,
-              public searchService: SearchService,
-              public overlayService: OverlayService,
-              public mapService: MapService,
-              public layerService: LayerService,
-              public dataSourceService: DataSourceService,
-              public contextService: ContextService,
-              public cdRef: ChangeDetectorRef) {}
+  constructor(
+    private route: ActivatedRoute,
+    public authService: AuthService,
+    public featureService: FeatureService,
+    public mediaService: MediaService,
+    public toolService: ToolService,
+    public searchService: SearchService,
+    public overlayService: OverlayService,
+    public mapService: MapService,
+    public layerService: LayerService,
+    public dataSourceService: DataSourceService,
+    public contextService: ContextService,
+    public cdRef: ChangeDetectorRef
+  ) {}
 
   ngOnInit() {
     window['IGO'] = this;
 
-    this.authService.authenticate$
-          .subscribe(() => this.contextLoaded = false);
+    this.authService.authenticate$.subscribe(
+      () => (this.contextLoaded = false)
+    );
 
-    this.features$$ = this.featureService.features$
-      .subscribe((features) => this.handleFeaturesChange(features));
+    this.features$$ = this.featureService.features$.subscribe(features =>
+      this.handleFeaturesChange(features)
+    );
 
-    this.selectedFeature$$ = this.featureService.selectedFeature$
-      .subscribe((feature) => this.handleFeatureSelect(feature));
+    this.selectedFeature$$ = this.featureService.selectedFeature$.subscribe(
+      feature => this.handleFeatureSelect(feature)
+    );
 
-    this.context$$ = this.contextService.context$
-      .subscribe((context) => this.handleContextChange(context));
+    this.context$$ = this.contextService.context$.subscribe(context =>
+      this.handleContextChange(context)
+    );
   }
 
   ngOnDestroy() {
@@ -85,8 +100,10 @@ export class PortalComponent implements OnInit, OnDestroy {
   closeSidenav() {
     this.sidenavOpened = false;
     this.toastOpened = false;
-    if (this.mediaService.media$.value === 'mobile' &&
-        this.featureService.focusedFeature$.value) {
+    if (
+      this.mediaService.media$.value === 'mobile' &&
+      this.featureService.focusedFeature$.value
+    ) {
       this.toastShown = true;
     }
   }
@@ -131,7 +148,9 @@ export class PortalComponent implements OnInit, OnDestroy {
     const focusedFeature = this.featureService.focusedFeature$.value;
 
     let index = featuresList.findIndex(f => f.id === focusedFeature.id);
-    if (index < 0) { return; }
+    if (index < 0) {
+      return;
+    }
 
     if (action === PortalComponent.SWIPE_ACTION.LEFT) {
       index += 1;
@@ -154,14 +173,15 @@ export class PortalComponent implements OnInit, OnDestroy {
   private handleFeaturesChange(features: Feature[]) {
     if (features.length > 0) {
       if (this.mediaService.media$.value === 'mobile') {
-        if (features[0].type.toString() === 'Feature' &&
-           (features[0].source !== 'Nominatim (OSM)' &&
-           features[0].source !== 'ICherche Québec')) {
-
-             this.featureService.selectFeature(features[0]);
-             this.overlayService.setFeatures([features[0]], 'zoom');
-             this.toastShown = true;
-             return;
+        if (
+          features[0].type.toString() === 'Feature' &&
+          (features[0].source !== 'Nominatim (OSM)' &&
+            features[0].source !== 'ICherche Québec')
+        ) {
+          this.featureService.selectFeature(features[0]);
+          this.overlayService.setFeatures([features[0]], 'zoom');
+          this.toastShown = true;
+          return;
         }
       }
 
@@ -198,23 +218,25 @@ export class PortalComponent implements OnInit, OnDestroy {
     });
   }
 
-
   private addLayerByName(url: string, name: string) {
     const properties = {
-      title: name,
-      type: 'wms',
-      format: 'wms',
+      type: 'wms' as any,
+      // format: 'wms',
       url: url,
       params: {
         layers: name
       }
     };
 
-    this.dataSourceService.createAsyncDataSource(properties).pipe(
-      debounceTime(100)
-    ).subscribe(dataSource =>  {
-        this.map.addLayer(
-          this.layerService.createLayer(dataSource, properties));
+    this.dataSourceService
+      .createAsyncDataSource(properties)
+      .pipe(debounceTime(100))
+      .subscribe(dataSource => {
+        const layerOptions = {
+          title: name,
+          source: dataSource
+        };
+        this.map.addLayer(this.layerService.createLayer(layerOptions));
       });
   }
 }
