@@ -1,27 +1,35 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
-import { IgoModule,
-         provideIChercheSearchSource,
-         provideNominatimSearchSource,
-         provideDataSourceSearchSource,
-         provideConfigOptions,
-         RouteService } from '@igo2/igo2';
+import {
+  provideConfigOptions,
+  IgoMessageModule,
+  RouteService
+} from '@igo2/core';
+import { IgoSpinnerModule, IgoStopPropagationModule } from '@igo2/common';
+import { IgoAuthModule } from '@igo2/auth';
+import {
+  provideIChercheSearchSource,
+  provideNominatimSearchSource,
+  provideDataSourceSearchSource
+} from '@igo2/geo';
 
 import { environment } from '../environments/environment';
 import { PortalModule } from './pages';
 import { AppComponent } from './app.component';
 
 @NgModule({
-  declarations: [
-    AppComponent
-  ],
+  declarations: [AppComponent],
   imports: [
     BrowserModule,
+    BrowserAnimationsModule,
     RouterModule.forRoot([]),
-    IgoModule.forRoot(),
-
+    IgoAuthModule,
+    IgoMessageModule,
+    IgoSpinnerModule,
+    IgoStopPropagationModule,
     PortalModule
   ],
   providers: [
@@ -36,4 +44,4 @@ import { AppComponent } from './app.component';
   ],
   bootstrap: [AppComponent]
 })
-export class AppModule { }
+export class AppModule {}
