@@ -1,22 +1,40 @@
 import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import {
+  MatTooltipModule,
+  MatIconModule,
+  MatButtonModule,
+  MatSidenavModule
+} from '@angular/material';
 
-import { SharedModule } from '../../shared';
+import { IgoCoreModule } from '@igo2/core';
+import {
+  IgoPanelModule,
+  IgoBackdropModule,
+  IgoFlexibleModule
+} from '@igo2/common';
+import { IgoGeoModule } from '@igo2/geo';
+import { IgoContextModule } from '@igo2/context';
+import { IgoToolsModule } from '@igo2/tools';
 
-import { SidenavComponent } from './sidenav';
 import { PortalComponent } from './portal.component';
-import { ToastComponent } from './toast/toast.component';
 
 @NgModule({
   imports: [
-    SharedModule
+    CommonModule,
+    MatTooltipModule,
+    MatButtonModule,
+    MatIconModule,
+    MatSidenavModule,
+    IgoCoreModule,
+    IgoPanelModule,
+    IgoBackdropModule,
+    IgoFlexibleModule,
+    IgoGeoModule,
+    IgoContextModule,
+    IgoToolsModule
   ],
-  exports: [
-    PortalComponent
-  ],
-  declarations: [
-    PortalComponent,
-    SidenavComponent,
-    ToastComponent
-  ]
+  exports: [PortalComponent],
+  declarations: [PortalComponent]
 })
 export class PortalModule {}
